backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation & Serialization
pydantic[email]==2.5.0
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# AI & Document Processing
google-generativeai==0.3.2
Pillow==10.1.0
pdf2image==1.16.3
PyMuPDF==1.23.8

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

# CORS & Middleware
starlette==0.27.0

# Date & Time handling
python-dateutil==2.8.2
